refactor: 지원서 질문 섹션 4개 컴포넌트를 QuestionSection 하나로 통합
- common/engineering/visualization/analysis question section을 QuestionSection으로 통합
- supportsTable prop으로 TABLE 타입 질문 지원 여부 제어 (engineering/visualization)
- nextLabel prop으로 제출 모드(footerSubmit) vs 네비게이션 모드(footerNav) 분기
- apply-page.tsx에서 track별 분기 렌더링 제거, QuestionSection 단일 사용으로 간소화
- 중복 CSS 제거하여 question-section.css.ts 하나로 통합
